@@ -1,7 +1,6 @@
 class AuthTokenIssuer < ApplicationService
-
   DEFAULT_JWT_EXPIRY = 10800
-  DEFAULT_JWT_ALGORITHM = 'HS256'
+  DEFAULT_JWT_ALGORITHM = "HS256"
 
   def initialize(user_id:)
     @user_id = user_id
@@ -19,7 +18,7 @@ class AuthTokenIssuer < ApplicationService
   end
 
   def expiry
-    length = ENV['JWT_EXPIRY']&.to_i || DEFAULT_JWT_EXPIRY
+    length = ENV["JWT_EXPIRY"]&.to_i || DEFAULT_JWT_EXPIRY
     Time.now.to_i + length
   end
 
@@ -30,7 +29,7 @@ class AuthTokenIssuer < ApplicationService
         exp: expiry
       },
       secret_jwt_key,
-      ENV['JWT_ALGORITHM'] || DEFAULT_JWT_ALGORITHM
+      ENV["JWT_ALGORITHM"] || DEFAULT_JWT_ALGORITHM
     )
   end
 end
