@@ -2,6 +2,8 @@ require_relative "boot"
 
 require "rails/all"
 
+require "sidekiq"
+
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
@@ -26,18 +28,13 @@ module RailsApiBase
 
     config.middleware.insert_before 0, Rack::Cors do
       allow do
-        origins '*' # TODO - UPDATE THIS TO RESTRICT REQUEST ORIGINS
-        resource '*', headers: :any, methods: [:get, :post]
+        origins "*" # TODO - UPDATE THIS TO RESTRICT REQUEST ORIGINS
+        resource "*", headers: :any, methods: [:get, :post]
       end
     end
 
     config.logger = ActiveSupport::Logger.new("log/#{Rails.env}.log")
 
-    config.before_configuration do
-      env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      YAML.load(File.open(env_file))&.each do |key, value|
-        ENV[key.to_s] = value.to_s
-      end if File.exists?(env_file)
-    end
+    config.autoload_paths += %W[#{config.root}/app/sidekiq]
   end
 end
