@@ -1,6 +1,7 @@
 FactoryBot.define do
   factory :user do
-    username { 'coolguy1000' }
-    password { 'super-$3cur3-PWD' }
+    id { SecureRandom.uuid }
+    username { "coolguy1000" }
+    password { "super-$3cur3-PWD" }
   end
 end
