@@ -1,11 +1,11 @@
 RSpec.describe ApplicationService do
-  subject { ApplicationService.new() }
+  subject { ApplicationService.new }
 
-  describe '#log' do
+  describe "#log" do
     let(:message) { "I should be logged!" }
     let(:legit_log_levels) { [:debug, :info, :warn, :error, :fatal, :unknown] }
 
-    it 'logs the provided message at the expected level' do
+    it "logs the provided message at the expected level" do
       legit_log_levels.each do |level|
         expected_message = "#{subject.class.name}: ".cyan + message
         expect(Rails.logger).to receive(level).with(expected_message)
@@ -14,10 +14,10 @@ RSpec.describe ApplicationService do
     end
   end
 
-  describe '#success' do
+  describe "#success" do
     let(:response_content) { "content" }
 
-    it 'logs a success message and returns a successful ServiceResponse object' do
+    it "logs a success message and returns a successful ServiceResponse object" do
       expect(subject).to receive(:log).with(:info, "Succeeded".green)
       response = subject.send(:success, response_content)
       expect(response.class).to eq(ServiceResponse)
@@ -26,10 +26,10 @@ RSpec.describe ApplicationService do
     end
   end
 
-  describe '#failure' do
+  describe "#failure" do
     let(:response_content) { "content" }
 
-    it 'logs a failure message and returns a failed ServiceResponse object' do
+    it "logs a failure message and returns a failed ServiceResponse object" do
       expect(subject).to receive(:log).with(:info, "Failed".red)
       response = subject.send(:failure, response_content)
       expect(response.class).to eq(ServiceResponse)
