@@ -1,5 +1,4 @@
 RSpec.describe AuthTokenIssuer do
-
   include ActiveSupport::Testing::TimeHelpers
 
   let(:user_id) do
@@ -7,24 +6,23 @@ RSpec.describe AuthTokenIssuer do
     user.id
   end
   let(:env_jwt_expiry) { 500 }
-  let(:env_jwt_algorithm) { 'HS512' }
+  let(:env_jwt_algorithm) { "HS512" }
   let(:default_jwt_expiry) { AuthTokenIssuer::DEFAULT_JWT_EXPIRY }
   let(:default_jwt_algorithm) { AuthTokenIssuer::DEFAULT_JWT_ALGORITHM }
 
   before :example do |test|
     unless test.metadata[:skip_env_stubs]
       stub_const(
-        'ENV', 
+        "ENV",
         ENV.to_hash.merge(
-          'JWT_EXPIRY' => env_jwt_expiry.to_s,
-          'JWT_ALGORITHM' => env_jwt_algorithm.to_s
+          "JWT_EXPIRY" => env_jwt_expiry.to_s,
+          "JWT_ALGORITHM" => env_jwt_algorithm.to_s
         )
       )
     end
   end
 
-  describe '#call' do
-
+  describe "#call" do
     subject { AuthTokenIssuer.call(user_id: user_id) }
 
     it "responds with a valid JWT auth token for the provided user" do
@@ -40,7 +38,7 @@ RSpec.describe AuthTokenIssuer do
     end
 
     context "when provided user does not exist" do
-      let(:user_id) { (User.last&.id || 0) + 1 }
+      let(:user_id) { SecureRandom.uuid } # repeat is EXTREMELY unlikely
 
       it "responds with a failed ServiceResponse" do
         response = subject
@@ -50,7 +48,6 @@ RSpec.describe AuthTokenIssuer do
     end
 
     context "when no ENV variables are set" do
-
       it "uses default values to produce a valid JWT auth token", skip_env_stubs: true do
         travel_to Time.now do
           response = subject
@@ -65,18 +62,17 @@ RSpec.describe AuthTokenIssuer do
     end
   end
 
-  describe '#expiry' do
-
+  describe "#expiry" do
     subject { AuthTokenIssuer.new(user_id: user_id) }
 
-    it 'provides a date time integer set forward the amount of time specified in ENV' do
+    it "provides a date time integer set forward the amount of time specified in ENV" do
       travel_to Time.now do
         expect(subject.send(:expiry)).to eq(Time.now.to_i + env_jwt_expiry)
       end
     end
 
     context "when no ENV variables are set" do
-      it 'provides a date time integer set forward with the default', skip_env_stubs: true do
+      it "provides a date time integer set forward with the default", skip_env_stubs: true do
         travel_to Time.now do
           expect(subject.send(:expiry)).to eq(Time.now.to_i + default_jwt_expiry)
         end
